Crate hotg_rune_runtime

The Rune Runtime.

Feature Flags

This crate makes use of cargo features to selectively enable or disable functionality.

The following cargo features are available:

builtins - (default) enable various builtin outputs and capabilities

tflite - (default) enable support for TensorFlow Lite models

wasm3 - enable the WASM3 engine

wasmer - enable the wasmer engine

Re-exports

pub extern crate wasm3;

pub extern crate wasmer;

Modules

builtins

Builtin modules.

models

Functions for handling various “well-known” model formats.

Structs

ModelMetadata

Metadata for a model node.

NodeMetadata

Metadata for a node in the ML pipeline, typically an input or output.

Runtime

A loaded Rune.

Tensor

A n-dimension array of numbers.

Enums

ElementType

The type of value that may be stored in a Tensor.

LoadError

OutputTensor

Traits

Model

An object that can do inference.

TensorElement

A numeric type that can be stored in a Tensor.