Module hotg_rune_runtime::builtins

Builtin modules.

Structs

AccelerometerSample

AccelerometerSamples

Arguments

Helper methods for reading arguments.

AudioClip

UnknownPixelFormat

Enums

AccelerometerParseError

Functions

accelerometer

Load an input tensor from a set of accelerometer samples.

image

Load an input tensor from an image, applying any transformations requested by the Rune.

random

raw

seeded_random

sound

Load an input from a sound clip, applying any transformations requested by the Rune.

source

Use the "source" argument to figure out which input to read.